[MIG][IMP][WIP] interclubs, scripts: Reflect interclub event color in calendar view.
In Odoo standard, it seems that we cannot set a dynamic background color
for a record in the calendar view, if the type a the color field is "#rrggbb".
This patch corrects this lack.

In top of the patch, a script is written and must be executed every time
the Odoo Standard repo has been pulled.
However this script is a WIP and cannot be used at this time.

As done in v13, this development has removed previous changes done in v12.