[MIG] interclubs: Update subject/body when changing template in IC event mail wizard.
Since v13.?, it seems related fields to composer (subject and body) are not changed
anymore when the composer itself has changed.
This commit force the update of those fields (for both "composer_id" and
"others_composer_id" composers).

On top of that, the method "onchange_template_id_wrapper" of the composer
does not clean the subject anymore when the associated template is empty.
This commit provides a manual cleaning of the subject when the template is unset.

Also clean-up old comment.