[FIX] interclubs: Open events view list when clicking on interclub kanban box.
Previously, an act_window was triggered when clicking on kanban box.
Code was confusing because the action had to be declared before the view,
otherwise the module could not be initialized.

With this commit, we call a python method which returns the action.
Code is more readable and no need to take care of the records declaration order.