[IMP] club: Add regeneration button for period alias email
When changing the `mail_local_part` of the config parameters, the email alias
of a period (which uses this parameter) is not updated.
This behavior is intended, because the email alias is writable by the user and
thus, should not be updated automatically.

This commit allows the user to easily regenerate the default email alias if
such parameter has changed (or for any other reason).

Other changes:
 - Let the "Emails" tab of a period be visible ponly in debug mode