[IMP] club: Add dynamic filter for periods and categories
In the `period` form view, it is possible to manage a `period.category` in
which a `category` can be selected.
Previously there was no domain on the categories to choose. Therefore it was
possible to select a category a second time (even if there is a constraint in
place).
This commit adds a domain (throught a new `category_ids` fields) in order to
have a dynamic domain on the category.

Other change:
 - The same mechanism has been implemented for the `category` form view.
   (Dynamic domain for the periods to choose.)