[FIX] club: Fix membership creation through alias email
On one hand, membership `period_id` field is required.
On the other hand, when creating/writing a membership record, `period_id` field
is removed from values [1] (among other things, it avoids to do an unnecessary
write on the `period.category` related record).
Thus, when creating a `membership` record outside of the form view (i.e. with
the email alias mechanism), no value was set for `period_id` field, which led
to an error.

Because `period_id` is related to `period_category_id` (which is required),
this field will always be set anyway.
Thus it is not needed to be required anymore (except in views).

[1] https://github.com/pieral85/sport_club_manager/blob/602b2791/club/models/membership.py#L354