[REF] club: Use standard auto-vacuum mechanism for periods update
Previously, a custom cron was running a server action in order to trigger the
recompute of the active/current/upcoming fields for a period.

With this commit, the `@api.autovacuum` decorator is used intead and the
method is automatically called w/ the standard `Base: Auto-vacuum internal
data` cron.

Other changes:
 - Add some logs during the update of the periods.