[FIX] club_account: Allow committee managers to modify period and categories
To create, write and delete periods, categories and period categories,
committee manager users must be able to have additionnal rights on some
standard product derivated models.

Note: It is not necessary to give read access right to the committee users
because this group already inherits from the `Internal User` one.