[FIX] interclubs: Select relevant mail template for Interclub Event Mail Wizard
Previously, when opening the "Interclub Event Mail Wizard", the
`calendar_template_meeting_invitation` was set as the default mail template.
With this commit, the "Interclub Event Opening/Confirmation/Cancellation" (*)
is choosen.

(*) depending on the "role" for which the wizard is opened (interclub event
"to_open"/"to_confirm"/"to_cancel" respectively).

Other change:
 - Set the domain as `interclub.event` for the mail templates to select in the
   wizard.