[FIX] interclubs: Avoid warning when preparing "Interclub Event Mail Wizard"
When setting key `src_model` in the dictionary going to call the action,
a `WARNING` is raised in the logs (*).
As this key did not seem to have any utility, it has been removed.

(*) WARNING [...] Action 'Interclub Event Mail Wizard' contains custom
properties 'src_model'. Passing them via the `params` or `context` properties
is recommended instead