[IMP] club: Let membership default email placeholders be "generic"
Because invitation and confirmation emails are rendered in mail composer with
'mass_mail` as `composition_mode', the placeholders are rendered with their
default values (but without the "Odoo blue text highlighting").

If default values seem too realistic (i.e. period is '2022-2023'), the user may
think this value will be actually be used when sending the email, which is
quite confusing.

Now with generic terms, the placeholders look like placeholders.

Other changes:
 - Fix typo
 - Remove extra blank line