[IMP] club, interclubs: Let mail template kind be more flexible
`kind` selection field has been added on a mail template instead of previous
booleans `is_membership_invitation_mail` and `is_membership_confirmation_mail`.
It allows to be more flexible in order to categorise the mail templates.
On top of that, it will be easier if new mail template categories are created
in the future.