[FIX] club: Allow mail composer to send multiple mails to the same address
On a partner, it happens to send a mass mail to multiple recipients.
On top of that, recipients could have the same email (i.e. 2 partners that have
the same `responsible_id`).
However in Odoo, if at least 2 mails are sent to the same partner (in the same
batch), one of them is cancelled (*).

With this commit, such usecase is allowed.

(*) https://github.com/odoo/odoo/blob/4d184bc6ca/addons/mail/wizard/mail_compose_message.py#L500-L502