[FIX] club, interclubs: Fix sending email issue for composer
In the `<mail.compose.message>.action_send_mail` method, we try to get the
closest lang for the records [1]. But because the `active_ids` are changed, the
`active_model` does not correspond to the ids anymore (records browsed are
completely wrong or inexistant).
This commit ensures both model and ids are consistant together.

Other changes:
 - Allow to get the closest lang for calendar attendee records (it is actually
   the closest lang of the partners).
   This is used when sending an email at the opening of an interlub event (w/
   mail template `interclubs.email_template_interclub_event_opening`).
 - `active_model` value passed in the membership context is not static anymore
   (doesn't change anything, but more beautiful)

[1] https://home.github.com/pieral85/sport_club_manager/blob/0683921ff7/club/models/mail.py#L83