[IMP] club: Add current (or upcoming) season as membership default value
Note:
In method `_modify_period_category_vals` of the `membership` model, `period_id`
and `category_id` keys were removed from the values used in the create/write
method. It had the advantage not to trigger the `write` method of the related
`period.category` model.
However, because this commit implements a default value/function for the period
of the membership, this value is called before saving the record, as
`period_id` has been previously removed from the values.
Thus, this commits reimplements both `period_id` and `category_id` keys, with
values coming from the `period_category_id`, so that they keep the same values.
Instead, it cancels the call of the `write` method on the `period.category` (as
both `period_id` and `category_id` field values as supposed to be the same).
This is not beautiful code, but it works...