[REF] club{_account}: Refactor dynamic window action view opening
This window action opens a view based on the number of record(s) to render.
Basically, the form view is rendered if there is only 1 record, otherwise, the
tree view is choosen.

As this feature was used in multiple places, it has now be refactored and
located in the `models.Model` class, so that it can be used anywhere.

Other change:
 - Slightly change the conditions fow which such views must be rendered, after
   mails are sent with a mail composer.