[IMP] club: Improve dates search on `period` model
Previously, it was allowed to search for a period based on the year.
i.e. searching "2023" (as a name) returned all periods for which at least one
     day (between start and end date) is in 2023.

But this functionality was done on the `name` field, which is not relevant.
With this commit, a new field `year` is created only for this purpose.