[ADD] club{_account}: Automatically archive membership when being rejected
It may be useful to let some memberships be not visible/searchable, specially
when they have just been rejected.
i.e. When sending mass mail on memberships, we usually do not want to include
rejected records (by default).
This is goal is achieved by archiving records being in this state.

This includes:
 - Add filters to target actived or archived memberships.

Other changes:
 - Let some membership fields be optional (tree view)
 - Fix a typo