[IMP] club: Allow to edit multiple membership prices at once
This may be useful after a smart duplication of a period occured, with new
prices being changed on the period categories, as newly duplicated memberships
still stick to the old prices.

A better solution (needing development) would be to allow the edition of the
new prices, directly in the period duplication wizard.