[IMP] club: Add age field on contact and membership models
This field is useful as we often need to know if a member is minor or not.

On the `res.partner` model, the age is computed, based on today's date.
On the `membership` model, the age is computed, based on first day of the
current season.

As a fallback value, the age is set to 0 if some information are missing
(contact has no birthdate, season is not set, ...).

This field has been implemented in following views:
 - `res.partner`: form and tree views (hidden by default in the latter one)
 - `membership`: kanban, form and tree views (hidden by default in the latter
    one)