[IMP] club: Allow to directly edit member information on membership view
When creating a membership, it often happens that the member also needs to be
created.
Previously, member information to fill had to be done by going to the
`res.partner` view, through the `member_id` field.

With this commit, all information (membership and member) can be filled on one
single view.

New changes done on `memberhip` model:
 - Add multiple fields related to current member (address, mail, phone/mobile,
   ...).
 - Add LFBB information related to current member.
 - Avoid to select contact person if it is the same as the member.

Other change:
 - When changing membership email, duplicate the email on the member itself, if
   not email was already set.
 - Add member image on membership form view.
 - Move age inside the membership form view.