[IMP] club_account: Allow to attach any membership on invoice
Previously, it was not possible to add or remove a membership on an account
move line (it had to be done from the membership itself).

In this situation, it was not possible to have memberships (*) set on the same
invoice. This is annoying as it sometimes happens that 2 people from the same
family pay together.
This commit adds this feature.

(*) It was actually possible but only if there was a dependance between those
memberships (e.g. parent with child(ren)).