[FIX] club_account: Fix membership price paid when no invoice attached
When searching searching value 0.0 as the price paid on a membership which is
not related to any invoice line, no record is returned.
This is because the computed price paid is considered as `False` by the ORM,
which is not correct. (The price displayed is actually 0.0 (which is correct),
but it is only because the `False` value is converted into the float 0.0.)

With this commit, we force the paid price to be 0.0 in such situation.