[IMP] club: Reset membership price at period smart duplication
For some memberships, the price due may diverge from the default one of the
current period category (for instance for a discount).
In this situation, when duplicating the related period, we don't want this
custom price anymore. Instead, we want to apply the default price on the newly
created membership.